+++ code_crafters_app/src/components/pages/HomePage/DiscountUserForm/index.jsx
@@ -1,7 +1,19 @@
+import React from 'react';
 import s from './DiscountUserForm.module.css';
 import SubmitForm from '../../../UI/SubmitForm/index';
+import { sendSaleRequest } from '../../../../redux/actions/postRequests';
 
 export default function DiscountUserForm() {
+	const handleSaleSubmit = async data => {
+		const saleData = {
+			name: data.name,
+			email: data.email,
+		};
+
+		const response = await sendSaleRequest(saleData);
+		return response;
+	};
+
 	return (
 		<div className={s.discount_user_form}>
 			<h2 className={s.first_order}>5% off on the first order</h2>
@@ -10,6 +22,7 @@ export default function DiscountUserForm() {
 					style={s}
 					button='Get a discount'
 					onSuccess='The discount has been successfully sent by email'
+					submitFunction={handleSaleSubmit} // Передача функции для отправки данных на скидку
 				/>
 			</div>
 		</div>
